feat: Move App.tsx to root folder for visibility
